Refactor & increased maintainability (#67)

+++ src-tauri/src/platform/shared.rs
@@ -1,5 +1,6 @@
-use raw_window_handle::{HasRawWindowHandle, RawWindowHandle};
 use std::time::Duration;
+
+use raw_window_handle::{HasRawWindowHandle, RawWindowHandle};
 use tauri::Window;
 
 use crate::data::Point;
@@ -18,7 +19,7 @@ pub const IGNORE_AREA_BOTTOM: u32 = 110;
 pub fn get_window_id(window: &Window) -> Option<u64> {
     #[allow(unused_variables)]
     match window.raw_window_handle() {
-        RawWindowHandle::Xlib(handle) => Some(handle.window as u64),
+        RawWindowHandle::Xlib(handle) => Some(handle.window),
         RawWindowHandle::Win32(handle) => Some(handle.hwnd as u64),
         RawWindowHandle::AppKit(handle) => {
             #[cfg(target_os = "macos")]
